Fix runtime type issue

+++ paq2.cpp
@@ -330,7 +330,7 @@ using std::swap;
 // 8-32 bit unsigned types, adjust as appropriate
 typedef unsigned char U8;
 typedef unsigned short U16;
-typedef unsigned long U32;
+typedef unsigned int U32;
 
 class U24 {  // 24-bit unsigned int
   U8 b0, b1, b2;  // Low, mid, high byte
