+++ paq5/paq5-emilcont-deuterium.cpp
@@ -468,7 +468,7 @@ template <class T> inline int size(const T& t) {return t.size();}
 // 8-32 bit unsigned types, adjust as appropriate
 typedef unsigned char U8;
 typedef unsigned short U16;
-typedef unsigned long U32;
+typedef unsigned int U32;
 
 // Fail if out of memory
 void handler() {
